Module aws_sdk_devopsguru::error

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

throttling_exception

See ThrottlingException

validation_exception

See ValidationException

Structs

AccessDeniedException

You don't have permissions to perform the requested operation. The user or role that is making the request must have at

AddNotificationChannelError

ConflictException

DescribeAccountHealthError

DescribeAccountOverviewError

DescribeAnomalyError

DescribeFeedbackError

DescribeInsightError

DescribeResourceCollectionHealthError

DescribeServiceIntegrationError

GetCostEstimationError

GetResourceCollectionError

InternalServerException

An internal failure in an Amazon service occurred.

ListAnomaliesForInsightError

ListEventsError

ListInsightsError

ListNotificationChannelsError

ListRecommendationsError

PutFeedbackError

RemoveNotificationChannelError

ResourceNotFoundException

A requested resource could not be found

SearchInsightsError

ServiceQuotaExceededException

The request contains a value that exceeds a maximum quota.

StartCostEstimationError

ThrottlingException

The request was denied due to a request throttling.

UpdateResourceCollectionError

UpdateServiceIntegrationError

ValidationException

Enums

AddNotificationChannelErrorKind

DescribeAccountHealthErrorKind

DescribeAccountOverviewErrorKind

DescribeAnomalyErrorKind

DescribeFeedbackErrorKind

DescribeInsightErrorKind

DescribeResourceCollectionHealthErrorKind

DescribeServiceIntegrationErrorKind

GetCostEstimationErrorKind

GetResourceCollectionErrorKind

ListAnomaliesForInsightErrorKind

ListEventsErrorKind

ListInsightsErrorKind

ListNotificationChannelsErrorKind

ListRecommendationsErrorKind

PutFeedbackErrorKind

RemoveNotificationChannelErrorKind

SearchInsightsErrorKind

StartCostEstimationErrorKind

UpdateResourceCollectionErrorKind

UpdateServiceIntegrationErrorKind