Module aws_sdk_devopsguru::model

Modules

anomaly_reported_time_range

See AnomalyReportedTimeRange

anomaly_source_details

See AnomalySourceDetails

anomaly_time_range

See AnomalyTimeRange

cloud_formation_collection

See CloudFormationCollection

cloud_formation_collection_filter

See CloudFormationCollectionFilter

cloud_formation_cost_estimation_resource_collection_filter

See CloudFormationCostEstimationResourceCollectionFilter

cloud_formation_health

See CloudFormationHealth

cloud_watch_metrics_detail

See CloudWatchMetricsDetail

cloud_watch_metrics_dimension

See CloudWatchMetricsDimension

cost_estimation_resource_collection_filter

See CostEstimationResourceCollectionFilter

cost_estimation_time_range

See CostEstimationTimeRange

end_time_range

See EndTimeRange

event

See Event

event_resource

See EventResource

event_time_range

See EventTimeRange

insight_feedback

See InsightFeedback

insight_health

See InsightHealth

insight_time_range

See InsightTimeRange

list_events_filters

See ListEventsFilters

list_insights_any_status_filter

See ListInsightsAnyStatusFilter

list_insights_closed_status_filter

See ListInsightsClosedStatusFilter

list_insights_ongoing_status_filter

See ListInsightsOngoingStatusFilter

list_insights_status_filter

See ListInsightsStatusFilter

notification_channel

See NotificationChannel

notification_channel_config

See NotificationChannelConfig

ops_center_integration

See OpsCenterIntegration

ops_center_integration_config

See OpsCenterIntegrationConfig

prediction_time_range

See PredictionTimeRange

proactive_anomaly

See ProactiveAnomaly

proactive_anomaly_summary

See ProactiveAnomalySummary

proactive_insight

See ProactiveInsight

proactive_insight_summary

See ProactiveInsightSummary

reactive_anomaly

See ReactiveAnomaly

reactive_anomaly_summary

See ReactiveAnomalySummary

reactive_insight

See ReactiveInsight

reactive_insight_summary

See ReactiveInsightSummary

recommendation

See Recommendation

recommendation_related_anomaly

See RecommendationRelatedAnomaly

recommendation_related_anomaly_resource

See RecommendationRelatedAnomalyResource

recommendation_related_anomaly_source_detail

See RecommendationRelatedAnomalySourceDetail

recommendation_related_cloud_watch_metrics_source_detail

See RecommendationRelatedCloudWatchMetricsSourceDetail

recommendation_related_event

See RecommendationRelatedEvent

recommendation_related_event_resource

See RecommendationRelatedEventResource

resource_collection

See ResourceCollection

resource_collection_filter

See ResourceCollectionFilter

search_insights_filters

See SearchInsightsFilters

service_collection

See ServiceCollection

service_health

See ServiceHealth

service_insight_health

See ServiceInsightHealth

service_integration_config

See ServiceIntegrationConfig

service_resource_cost

See ServiceResourceCost

sns_channel_config

See SnsChannelConfig

start_time_range

See StartTimeRange

update_cloud_formation_collection_filter

See UpdateCloudFormationCollectionFilter

update_resource_collection_filter

See UpdateResourceCollectionFilter

update_service_integration_config

See UpdateServiceIntegrationConfig

validation_exception_field

See ValidationExceptionField

Structs

AnomalyReportedTimeRange

AnomalySourceDetails

Details about the source of the anomalous operational data that triggered the anomaly. The

AnomalyTimeRange

CloudFormationCollection

CloudFormationCollectionFilter

CloudFormationCostEstimationResourceCollectionFilter

Information about an AWS CloudFormation stack used to create a monthly cost estimate for DevOps Guru to analyze

CloudFormationHealth

CloudWatchMetricsDetail

CloudWatchMetricsDimension

The dimension of a Amazon CloudWatch metric that is used when DevOps Guru analyzes the resources in your

CostEstimationResourceCollectionFilter

Information about a filter used to specify which AWS resources are analyzed to create

CostEstimationTimeRange

The time range of a cost estimation.

EndTimeRange

Event

EventResource

EventTimeRange

InsightFeedback

InsightHealth

InsightTimeRange

A time ranged that specifies when the observed behavior in an insight started and

ListEventsFilters

ListInsightsAnyStatusFilter

ListInsightsClosedStatusFilter

ListInsightsOngoingStatusFilter

ListInsightsStatusFilter

NotificationChannel

NotificationChannelConfig

Information about notification channels you have configured with DevOps Guru.

OpsCenterIntegration

OpsCenterIntegrationConfig

PredictionTimeRange

ProactiveAnomaly

Information about an anomaly. This object is returned by ListAnomalies.

ProactiveAnomalySummary

Details about a proactive anomaly. This object is returned by

ProactiveInsight

Details about a proactive insight. This object is returned by ListInsights.

ProactiveInsightSummary

Details about a proactive insight. This object is returned by

ReactiveAnomaly

Details about a reactive anomaly. This object is returned by ListAnomalies.

ReactiveAnomalySummary

Details about a reactive anomaly. This object is returned by

ReactiveInsight

ReactiveInsightSummary

Information about a reactive insight. This object is returned by

Recommendation

Recommendation information to help you remediate detected anomalous behavior that

RecommendationRelatedAnomaly

RecommendationRelatedAnomalyResource

RecommendationRelatedAnomalySourceDetail

RecommendationRelatedCloudWatchMetricsSourceDetail

RecommendationRelatedEvent

RecommendationRelatedEventResource

ResourceCollection

ResourceCollectionFilter

SearchInsightsFilters

ServiceCollection

A collection of the names of AWS services.

ServiceHealth

Represents the health of an AWS service.

ServiceInsightHealth

Contains the number of open proactive and reactive insights in an analyzed AWS service.

ServiceIntegrationConfig

ServiceResourceCost

An object that contains information about the estimated monthly cost to

SnsChannelConfig

Contains the Amazon Resource Name (ARN) of an Amazon Simple Notification Service topic.

StartTimeRange

UpdateCloudFormationCollectionFilter

Contains the names of AWS CloudFormation stacks used to update a collection of stacks. You can specify up to 500 AWS CloudFormation stacks.

UpdateResourceCollectionFilter

UpdateServiceIntegrationConfig

ValidationExceptionField

Enums

AnomalySeverity

AnomalyStatus

CloudWatchMetricsStat

CostEstimationServiceResourceState

CostEstimationStatus

EventClass

EventDataSource

InsightFeedbackOption

InsightSeverity

InsightStatus

InsightType

Locale

OptInStatus

ResourceCollectionType

ServiceName

UpdateResourceCollectionAction

ValidationExceptionReason